+++ optimize.py
@@ -56,9 +56,9 @@ def calitems():
 # Worker
 def test(calitems):
     t1 = datetime.now()
-    ids = cal.add_items(items=calitems)
+    ids = cal.bulk_create(items=calitems)
     t2 = datetime.now()
-    cal.delete_items(ids)
+    cal.bulk_delete(ids)
     t3 = datetime.now()
 
     delta1 = t2 - t1